pyproject.toml
[tool.poetry]
name = "defectio"
version = "0.3.0"
description = "Wrapper for Revolt API"
authors = ["Leon Bowie <leon@bowie-co.nz>"]
license = "MIT"
documentation = "https://defectio.readthedocs.io"
homepage = "https://github.com/Darkflame72/defectio"
readme = "README.md"
repository = "https://github.com/Darkflame72/defectio"
classifiers = [
  "Natural Language :: English",
  "Framework :: AsyncIO",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Utilities",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
]
include = ["defectio/py.typed"]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.7.4"
ulid-py = "^1.1.0"
msgpack = "^1.0.2"
aiofiles = "^0.7.0"
aiodns = {version = "^3.0.0", extras = ["speedups"]}
Brotli = {version = "^1.0.9", extras = ["speedups"]}
cchardet = {version = "^2.1.7", extras = ["speedups"]}
attrs = "^21.4.0"


[tool.poetry.dev-dependencies]
black = "^21.9b0"
mypy = "^0.910"
flake8 = "^3.9.2"
nox = "^2021.6.12"
darglint = "^1.8.0"
nox-poetry = "^0.8.6"
coverage = {extras = ["toml"], version = "^5.5"}
pytest = "^6.2.5"
pytest-asyncio = "^0.15.1"
Sphinx = "^4.2.0"
sphinxcontrib-trio = "^1.1.2"
pep8-naming = "^0.12.1"
Pygments = "^2.10.0"
flake8-annotations = "^2.6.2"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.3"
flake8-bugbear = "^21.9.2"
flake8-docstrings = "^1.6.0"
flake8-import-order = "^0.18.1"
pre-commit = "^2.15.0"
pre-commit-hooks = "^4.0.1"
safety = "^1.10.3"
reorder-python-imports = "^2.6.0"
aioresponses = "^0.7.2"
furo = "^2021.9.22"
sphinxext-opengraph = "^0.4.2"
sphinx-copybutton = "^0.4.0"
sphinx-autobuild = "^2021.3.14"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
